Adición de modelos 3D plantas vs zombies clasicos, correción de bug dentro de los soles 3D y mejoras visuales Beta 12.3.2

+++ Assets/Scripts/GameManager.cs
@@ -37,7 +37,7 @@ public class GameManager : MonoBehaviour
         Vector2 posifinal = new Vector2(posiX, posiY);
         return posifinal;
     }
-    public void AñadirSoles(ushort cantidad)
+    public void AñadirSoles(uint cantidad)
     {
         if (recogidaSol.enabled == false) StartCoroutine(SonidoSol());
         else StartCoroutine(SonidoSolDos());
